move cli into own package
