+++ kitsune-cli/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "kitsune-cli"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+chrono = "0.4.23"
+clap = { version = "4.1.4", features = ["derive"] }
+dotenvy = "0.15.6"
+envy = "0.4.2"
+kitsune-db = { version = "0.1.0", path = "../kitsune-db" }
+sea-orm = { version = "0.10.7", features = ["sqlx-postgres", "runtime-tokio-rustls"] }
+serde = { version = "1.0.152", features = ["derive"] }
+tokio = { version = "1.25.0", features = ["full"] }
+tracing-subscriber = "0.3.16"
+uuid = "1.3.0"
