+++ kitsune-cli/src/config.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct Configuration {
+    pub database_url: String,
+}
